requirements.txt
flask
requests
google-genai
flask-cors
sendgrid
